// src/features/pdf/styles/invoice.styles.d.ts
export declare const styles: {
  page: {
    fontSize: number;
    letterSpacing: number;
  };
  container: {
    paddingVertical: number;
    paddingHorizontal: number;
  };
  header: {
    flexDirection: 'row';
    alignItems: 'center';
    justifyContent: 'space-between';
    backgroundColor: string;
    color: string;
    borderBottomWidth: number;
    borderBottomColor: string;
  };
  companyBranding: {
    flexDirection: 'row';
    alignItems: 'center';
    gap: number;
  };
  logoContainer: {
    width: number;
  };
  logo: {
    objectFit: string;
    objectPosition: string;
    width: string;
    height: string;
  };
  companyName: {
    fontSize: number;
    fontWeight: string;
  };
  invoicePDFHeader: {
    gap: number;
    alignItems: 'flex-end';
  };
  invoiceTitle: {
    color: string;
  };
  invoiceNumber: {
    fontSize: number;
  };
  invoiceDate: {
    color: string;
  };
  body: {
    gap: number;
  };
  paymentSummary: {
    flexDirection: 'row';
    justifyContent: 'space-between';
    alignItems: 'center';
  };
  payeeSection: {
    gap: number;
  };
  totalSectionLabelBox: {
    gap: number;
  };
  payeeSectionLabel: {
    color: string;
  };
  payeeSectionName: {
    fontSize: number;
  };
  totalSection: {
    color: string;
    alignItems: 'center';
    borderWidth: number;
    borderColor: string;
    backgroundColor: string;
    paddingVertical: number;
    paddingHorizontal: number;
    gap: number;
    borderRadius: number;
  };
  goldUnderline: {
    borderBottomWidth: number;
    borderBottomColor: string;
    width: number;
    marginTop: number;
    alignSelf: 'flex-start';
  };
  totalSectionAmount: {
    fontSize: number;
    fontWeight: string;
  };
  tableSection: {
    borderRadius: number;
  };
  tableHeader: {
    flexDirection: 'row';
    backgroundColor: string;
    paddingVertical: number;
    paddingHorizontal: number;
    fontWeight: string;
    color: string;
    borderTopLeftRadius: number;
    borderTopRightRadius: number;
    width: string;
  };
  tableCellCol1: {
    flex: number;
  };
  tableCellCol2: {
    flex: number;
    textAlign: 'center';
  };
  tableCellCol3: {
    flex: number;
    textAlign: 'center';
  };
  tableCellCol4: {
    flex: number;
    textAlign: 'right';
  };
  tableBody: {
    borderLeftWidth: number;
    borderRightWidth: number;
    borderBottomWidth: number;
    borderColor: string;
    backgroundColor: string;
    borderBottomLeftRadius: number;
    borderBottomRightRadius: number;
    width: string;
  };
  tableBodyRow: {
    paddingHorizontal: number;
  };
  tableBodyRowInner: {
    flexDirection: 'row';
    borderBottom: string;
    paddingVertical: number;
  };
  tableBodyRowTotal: {
    flexDirection: 'row';
    justifyContent: 'space-between';
    alignItems: 'center';
    paddingVertical: number;
    borderTop: string;
    color: string;
    fontSize: number;
  };
  tableBodyRowTotalAmount: {
    fontSize: number;
  };
  additionalInfoSection: {
    flexDirection: 'row';
  };
  additionalInfoColumns: {
    flexDirection: 'row';
    width: string;
    gap: number;
  };
  additionalInfoColumnsContent: {
    width: string;
    gap: number;
    fontSize: number;
  };
  additionalInfoColumnsTitle: {
    color: string;
    fontWeight: string;
  };
  additionalInfoColumnsText: {
    lineHeight: number;
  };
  footerSection: {
    marginTop: string;
  };
  footerSectionContent: {
    paddingHorizontal: number;
  };
  footerDividerSection: {
    borderTop: string;
    paddingVertical: number;
    alignItems: 'center';
    gap: number;
  };
  footerSectionAgencyName: {
    color: string;
  };
};